Rename ScriptRule -> ScriptExtension
