+++ spin/core/src/test/java/org/operaton/spin/impl/test/Script.java
@@ -22,13 +22,13 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 /**
- * Method annotation to signal a script test for the {@link ScriptRule} {@link org.junit.jupiter.api.extension.Extension}.
+ * Method annotation to signal a script test for the {@link ScriptExtension} {@link org.junit.jupiter.api.extension.Extension}.
  * The name of the script is optional and if omitted the package name is used as directory
  * and the ClassName.MethodName.ScriptExtension as filename. Additionally a list of
  * variables can be defined and the execute flag decides whether the script is execute
  * before the test function is called.
  *
- * The script variables can be accessed trough the {@link ScriptRule#variables}
+ * The script variables can be accessed trough the {@link ScriptExtension#variables}
  * field.
  *
  * Example usage:
