+++ spin/core/src/test/java/org/operaton/spin/impl/test/script/PythonScriptExtensionTest.java
@@ -17,11 +17,11 @@
 package org.operaton.spin.impl.test.script;
 
 import org.operaton.spin.impl.test.ScriptEngine;
-import org.operaton.spin.impl.test.ScriptRuleTest;
+import org.operaton.spin.impl.test.ScriptExtensionTest;
 
 /**
  * @author Sebastian Menski
  */
-@ScriptEngine("graal.js")
-public class JavascriptScriptRuleTest extends ScriptRuleTest {
+@ScriptEngine("python")
+public class PythonScriptExtensionTest extends ScriptExtensionTest {
 }
