+++ spin/core/src/test/resources/org/operaton/spin/impl/test/script/JavascriptScriptExtensionNashornTest.shouldSetFoo.js
