+++ spin/core/src/test/resources/org/operaton/spin/impl/test/script/JavascriptScriptExtensionTest.shouldSetBar.js
