+++ spin/core/src/test/resources/org/operaton/spin/impl/test/script/JavascriptScriptExtensionTest.shouldSetFoo.js
