+++ spin/core/src/test/resources/org/operaton/spin/impl/test/script/PythonScriptExtensionTest.shouldSetFoo.py
