+++ spin/core/src/test/resources/org/operaton/spin/impl/test/script/RubyScriptExtensionTest.shouldSetBar.rb
