+++ spin/core/src/test/resources/org/operaton/spin/impl/test/script/RubyScriptExtensionTest.shouldSetFoo.rb
